Create dynamic functionality for container and box sizing
Create variable size that is equal to 100 / number
of boxes that the user enters. This number is then
calculated and used to set each box's height and
width to a percentage so that a grid is always
formed. The sizing was removed from the css file
so that the js could dynamically update sizing
using this methodology.