Create getRandomColor function
Create getRandomColor function to randomly
generate a hexadecimal color and return that color.

Revoved :hover pseudo class from css. The hover
functionality is now built into the js using
event listeners. Moverover event listener was
added to myDivs and uses a function to set background color
to the randomly generated color from random color
function. The mouserout event listener is added to
myDivs so that the color is set to blank
when mouse leaves each div.